+++ service_catalog/api/serializers/request_serializers.py
@@ -133,10 +133,9 @@ class OperationRequestSerializer(ModelSerializer):
 
     def validate(self, data):
         super(OperationRequestSerializer, self).validate(data)
-
-        if self.operation.is_admin_operation and not self.user.has_perm("service_catalog.admin_request_on_instance"):
+        if not self.user.has_perm("service_catalog.request_on_instance"):
             raise PermissionDenied
-        if not self.operation.is_admin_operation and not self.user.has_perm("service_catalog.request_on_instance"):
+        if not self.user.has_perm(self.operation.permission.permission_str, self.squest_instance):
             raise PermissionDenied
         if self.squest_instance.state not in [InstanceState.AVAILABLE]:
             raise PermissionDenied("Instance not available")
@@ -146,6 +145,8 @@ class OperationRequestSerializer(ModelSerializer):
             raise PermissionDenied("Operation service and instance service doesn't match")
         if self.operation.type not in [OperationType.UPDATE, OperationType.DELETE]:
             raise PermissionDenied("Operation type UPDATE and DELETE only")
+        if not self.operation.when_instance_authorized(self.squest_instance):
+            raise PermissionDenied("Operation not allowed")
         fill_in_survey = data.get("fill_in_survey")
         request_comment = data.get("request_comment")
         fill_in_survey.update({"request_comment": request_comment})