+++ service_catalog/models/portfolio.py
@@ -1,8 +1,12 @@
+import logging
+
 from django.db.models import ForeignKey, SET_NULL, CharField, ImageField
 from django.urls import reverse_lazy
 
 from Squest.utils.squest_model import SquestModel
+from profiles.models import Permission
 
+logger = logging.getLogger(__name__)
 
 class Portfolio(SquestModel):
     name = CharField(max_length=100)
@@ -34,3 +38,14 @@ class Portfolio(SquestModel):
         self.portfolio_list.update(**{"parent_portfolio": self.parent_portfolio})
         self.service_list.update(**{"parent_portfolio": self.parent_portfolio})
         super(Portfolio, self).delete(using, keep_parents)
+
+    def bulk_set_permission_on_operation(self, target_permission):
+        from service_catalog.models import Operation
+        logger.debug(f"Bulk edit permission on portfolio {self.name} to permission {target_permission}")
+        # check if all permission are already set to the target perm
+        all_permission_current_portfolio = Permission.objects.filter(operation__service__parent_portfolio=self).distinct()
+        # if the target perm is already the one used we do nothing
+        if all_permission_current_portfolio.count() == 1:
+            if all_permission_current_portfolio.first() == target_permission:
+                return
+        Operation.objects.filter(service__parent_portfolio=self).update(permission=target_permission)