+++ tests/test_service_catalog/test_api/test_urls/test_instance/test_crud.py
@@ -1,4 +1,4 @@
-from service_catalog.models import CustomLink, InstanceState
+from service_catalog.models import InstanceState
 from tests.permission_endpoint import TestPermissionEndpoint, TestingGetContextView, TestingPostContextView, \
     TestingPatchContextView, TestingPutContextView, TestingDeleteContextView
 from tests.test_service_catalog.base_test_request import BaseTestRequestAPI
@@ -12,19 +12,19 @@ class TestServiceCatalogInstanceCRUDPermissionsEndpoint(BaseTestRequestAPI, Test
         self.test_instance_2.service = self.update_operation_test_2.service
         self.test_instance.save()
         self.test_instance_2.save()
-
-        self.update_operation_test_2.is_admin_operation = True
+        self.update_operation_test_2.permission = self.admin_operation
         self.update_operation_test_2.save()
+
 
     def test_instance_views(self):
         testing_view_list = [
             TestingGetContextView(
                 url='api_instance_list_create',
-                perm_str='service_catalog.list_instance',
+                perm_str_list=['service_catalog.list_instance'],
             ),
             TestingPostContextView(
                 url='api_instance_list_create',
-                perm_str='service_catalog.add_instance',
+                perm_str_list=['service_catalog.add_instance'],
                 data={
                     "name": "New instance",
                     "service": self.service_test_2.id,
@@ -39,14 +39,14 @@ class TestServiceCatalogInstanceCRUDPermissionsEndpoint(BaseTestRequestAPI, Test
             ),
             TestingGetContextView(
                 url='api_operation_request_create',
-                perm_str='service_catalog.view_request',
+                perm_str_list=['service_catalog.view_request'],
                 url_kwargs={'instance_id': self.test_instance.id, 'operation_id': self.update_operation_test.id},
                 expected_status_code=405,
                 expected_not_allowed_status_code=405
             ),
             TestingPostContextView(
                 url='api_operation_request_create',
-                perm_str='service_catalog.request_on_instance',
+                perm_str_list=['service_catalog.request_on_instance', self.update_operation_test.permission.permission_str],
                 url_kwargs={'instance_id': self.test_instance.id, 'operation_id': self.update_operation_test.id},
                 data={
                     'fill_in_survey': {
@@ -55,14 +55,14 @@ class TestServiceCatalogInstanceCRUDPermissionsEndpoint(BaseTestRequestAPI, Test
             ),
             TestingGetContextView(
                 url='api_operation_request_create',
-                perm_str='service_catalog.view_request',
+                perm_str_list=['service_catalog.view_request'],
                 url_kwargs={'instance_id': self.test_instance_2.id, 'operation_id': self.update_operation_test_2.id},
                 expected_status_code=405,
                 expected_not_allowed_status_code=405
             ),
             TestingPostContextView(
                 url='api_operation_request_create',
-                perm_str='service_catalog.admin_request_on_instance',
+                perm_str_list=['service_catalog.request_on_instance', self.update_operation_test_2.permission.permission_str],
                 url_kwargs={'instance_id': self.test_instance_2.id, 'operation_id': self.update_operation_test_2.id},
                 data={
                     'fill_in_survey': {
@@ -72,12 +72,12 @@ class TestServiceCatalogInstanceCRUDPermissionsEndpoint(BaseTestRequestAPI, Test
             ),
             TestingGetContextView(
                 url='api_instance_details',
-                perm_str='service_catalog.view_instance',
+                perm_str_list=['service_catalog.view_instance'],
                 url_kwargs={'pk': self.test_instance.id}
             ),
             TestingPutContextView(
                 url='api_instance_details',
-                perm_str='service_catalog.change_instance',
+                perm_str_list=['service_catalog.change_instance'],
                 data={
                     'name': 'Instance PUT',
                     "service": self.service_test_2.id,
@@ -93,7 +93,7 @@ class TestServiceCatalogInstanceCRUDPermissionsEndpoint(BaseTestRequestAPI, Test
             ),
             TestingPatchContextView(
                 url='api_instance_details',
-                perm_str='service_catalog.change_instance',
+                perm_str_list=['service_catalog.change_instance'],
                 data={
                     'name': 'Instance PATCH',
                 },
@@ -101,7 +101,7 @@ class TestServiceCatalogInstanceCRUDPermissionsEndpoint(BaseTestRequestAPI, Test
             ),
             TestingDeleteContextView(
                 url='api_instance_details',
-                perm_str='service_catalog.delete_instance',
+                perm_str_list=['service_catalog.delete_instance'],
                 url_kwargs={'pk': self.test_instance.id}
             )
         ]