+++ tests/test_service_catalog/test_views/test_customer/test_catalog/test_catalog.py
@@ -1,6 +1,6 @@
 from django.urls import reverse
 
-from service_catalog.models import Request, Service, RequestMessage, Portfolio, Operation
+from service_catalog.models import Request, Service, RequestMessage, Portfolio, Operation, OperationType
 from tests.test_service_catalog.base_test_request import BaseTestRequest
 
 
@@ -9,6 +9,7 @@ class TestCustomerCatalogViews(BaseTestRequest):
     def setUp(self):
         super(TestCustomerCatalogViews, self).setUp()
         self.client.login(username=self.standard_user, password=self.common_password)
+
 
     def test_customer_list_service_in_root(self):
         url = reverse('service_catalog:service_catalog_list')
@@ -122,11 +123,18 @@ class TestCustomerCatalogViews(BaseTestRequest):
         self.assertEqual(created_request.comments.first().sender, self.standard_user)
 
     def test_customer_cannot_create_request_on_admin_only_operation(self):
+
         admin_create_operation = Operation.objects.create(name="create admin",
                                                           service=self.service_test,
                                                           job_template=self.job_template_test,
-                                                          is_admin_operation=True,
-                                                          process_timeout_second=20)
+                                                          process_timeout_second=20,
+                                                          permission=self.admin_operation)
+        # add a second create operation to avoid being redirected directly to the unique operation form
+        Operation.objects.create(name="create non admin",
+                                 service=self.service_test,
+                                 job_template=self.job_template_test,
+                                 type=OperationType.CREATE,
+                                 process_timeout_second=20)
         self.client.force_login(user=self.standard_user)
         args = {
             "service_id": self.service_test.id
@@ -153,3 +161,32 @@ class TestCustomerCatalogViews(BaseTestRequest):
         for step, data_step in enumerate(STEPS_DATA, 1):
             response = self.client.post(url, data=data_step)
             self.assertEqual(403, response.status_code)
+
+    def test_user_cannot_list_admin_create_operation(self):
+
+        Operation.objects.create(name="second create admin",
+                                 service=self.service_test,
+                                 job_template=self.job_template_test,
+                                 permission=self.admin_operation,
+                                 type=OperationType.CREATE,
+                                 process_timeout_second=20)
+        Operation.objects.create(name="third create non admin",
+                                 service=self.service_test,
+                                 job_template=self.job_template_test,
+                                 type=OperationType.CREATE,
+                                 process_timeout_second=20)
+        self.client.force_login(user=self.superuser)
+        args = {
+            "service_id": self.service_test.id
+        }
+        url = reverse('service_catalog:create_operation_list', kwargs=args)
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200)
+        # admin is able to get the 3 create operation
+        self.assertEqual(response.context_data["operation_list"].count(), 3)
+
+        self.client.force_login(user=self.standard_user)
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200)
+        # standard user get all create operation except the admin protected one
+        self.assertEqual(response.context_data["operation_list"].count(), 2)